add error module

+++ include/siphon/path.h
@@ -23,7 +23,7 @@ sp_path_split (SpRange16 *a, SpRange16 *b, const char *path, uint16_t plen, int 
 SP_EXPORT void
 sp_path_splitext (SpRange16 *a, SpRange16 *b, const char *path, uint16_t plen);
 
-SP_EXPORT uint16_t
+SP_EXPORT int
 sp_path_join (char *out, size_t len,
 		const char *a, uint16_t alen,
 		const char *b, uint16_t blen,
@@ -35,10 +35,10 @@ sp_path_clean (char *path, uint16_t len, SpPathMode mode);
 SP_EXPORT bool
 sp_path_match (const char *path, const char *match);
 
-SP_EXPORT ssize_t
+SP_EXPORT int
 sp_path_proc (char *buf, size_t buflen);
 
-SP_EXPORT ssize_t
+SP_EXPORT int
 sp_path_env (const char *name, char *buf, size_t buflen);
 
 #endif
